Almost done!
I cleaned up the front end, and fixed the listNGrandparents/children()
function. As a nice addition, I whipped up a fixGeneration() function,
which takes an int and returns the appropriate generation; this makes
the output much nicer. All that's really needed now is an amended
listSiblings() function, with half/step-siblings, and all that jazz.